+++ src/schemas/core/index.ts
@@ -6,3 +6,5 @@ export * from './aggregation.schema';
 export * from './callback.schema';
 export * from './link.schema';
 export * from './file.schema';
+export * from './polygonParts.schema';
+export * from './tile.schema';