+++ src/types/core/index.ts
@@ -6,3 +6,5 @@ export * from './part.type';
 export * from './link.type';
 export * from './file.type';
 export * from './callback.type';
+export * from './polygonParts.type';
+export * from './tile.type';